refactor: split CLI index.ts into command and utility modules
Break the 823-line monolithic index.ts into focused modules under
commands/ and utils/ directories for improved readability and
testability. No external behavior changes — all 60 existing tests
pass unchanged.
